Refactors mousedown event handler on canvas.
jQuery handles all browser compatibility problems regarding events quite well
and presents the x and y coordinates of the mouse pointer in e.pageX and
e.pageY. The offset of the canvas can be retrieved using jQuery as
canvas.offset(). The mousedown event fires if the mouse button is pressed.
A click event fires if a mouse button is pressed and released on the same
element.